+++ src/cmark_readme.rs
@@ -4,6 +4,7 @@
 use std::path::Path;
 use std::sync::Arc;
 
+use pulldown_cmark::Event;
 #[cfg(feature = "thiserror")]
 use thiserror::Error;
 
@@ -93,6 +94,11 @@ impl<'a, P, M> CMarkReadme<P, M> {
         &self.data
     }
 
+    /// Consumes the `CMarkReadme`, returning `CMarkData`.
+    pub fn into_data(self) -> CMarkData {
+        self.data
+    }
+
     /// Returns the package path.
     pub fn package_path(&self) -> &P {
         &self.package_path
@@ -106,6 +112,11 @@ impl<'a, P, M> CMarkReadme<P, M> {
     /// Iterate over `CMarkItem`s.
     pub fn iter(&self) -> CMarkDataIter<'_> {
         self.data.iter()
+    }
+
+    /// Iterate over pulldown-cmark events.
+    pub fn iter_events(&self) -> impl Iterator<Item = &Event<'_>> {
+        self.data.iter().filter_map(|item| item.event())
     }
 
     fn map<F>(mut self, func: F) -> CMarkReadme<P, M>