+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `CMarkDocs::map_links` function to fix links including intra-doc links.
+- `{CMarkDocs,CMarkReadme}::into_data` and `CMarkData::{into_items,from_items}`.
 
 ### Fixed
 - Fix url with scheme check